requirements.txt
docker-py==1.7.2
pytz==2013.8
requests==2.1.0
six==1.10.0
websocket-client==0.35.0
python-etcd==0.4.3
